add some alerts and messages

+++ client/src/components/customerTable/customerTable.js
@@ -8,6 +8,15 @@ import '../../css/it19951386.css';
 export default function CustomerTable() {
     let [customers, setCustomers] = useState([]);
     let [search, setSearch] = useState("");
+    const[cusAddMsg] = useState(localStorage.getItem("cusAddMsg"));//this sets logout msg if any
+    const[cusDelMsg] = useState(localStorage.getItem("cusDelMsg"));//this sets logout msg if any
+    const[cusUpMsg] = useState(localStorage.getItem("cusUpMsg"));//this sets logout msg if any
+
+    window.onload = function() {
+        localStorage.removeItem("cusAddMsg");
+        localStorage.removeItem("cusDelMsg");
+        localStorage.removeItem("cusUpMsg");
+    }//this resets the state on refreshing
 
     if(search.length > 0){
         customers = customers.filter((i) => {
@@ -24,9 +33,14 @@ export default function CustomerTable() {
     },[])//fetching customer data
     
     const deleteCustomer = async(id,cusID) => {
-        alert("Are sure want to delete " + cusID + " ?")
-        const deletion = await axios.delete(`http://localhost:5000/customers/delete/${id}`);
+        let deletion;
+        // alert("Are sure want to delete " + cusID + " ?")
+        if(window.confirm("Are you sure about deleting " + cusID + " ?")){
+            deletion = await axios.delete(`http://localhost:5000/customers/delete/${id}`);
+        }
+        //const deletion = await axios.delete(`http://localhost:5000/customers/delete/${id}`);
         if(deletion) {
+            localStorage.setItem('cusDelMsg', "Customer record of " + cusID +", was deleted successfully!");
             window.location = "/customers"
         }
          
@@ -80,6 +94,18 @@ export default function CustomerTable() {
         <path d="M8 2a5.53 5.53 0 0 0-3.594 1.342c-.766.66-1.321 1.52-1.464 2.383C1.266 6.095 0 7.555 0 9.318 0 11.366 1.708 13 3.781 13h8.906C14.502 13 16 11.57 16 9.773c0-1.636-1.242-2.969-2.834-3.194C12.923 3.999 10.69 2 8 2zm2.354 6.854l-2 2a.5.5 0 0 1-.708 0l-2-2a.5.5 0 1 1 .708-.708L7.5 9.293V5.5a.5.5 0 0 1 1 0v3.793l1.146-1.147a.5.5 0 0 1 .708.708z"/>
         </svg><span className="spinner-border spinner-border-sm" id="loading" role="status" aria-hidden="true" style={{display:'none'}}></span> Download</button>
         <input id="search-filter" className="form-control" type="text" placeholder="Search by name or ID..." onChange={(e) => {setSearch(e.target.value)}} value={search}/>
+        {cusAddMsg ? 
+        <div class="alert alert-success" role="alert">
+        {cusAddMsg}
+        </div> : null}
+        {cusDelMsg ? 
+        <div class="alert alert-success" role="alert">
+        {cusDelMsg}
+        </div> : null}
+        {cusUpMsg ? 
+        <div class="alert alert-success" role="alert">
+        {cusUpMsg}
+        </div> : null}
         <table id="table">
             <thead>
                 <tr >
@@ -101,10 +127,10 @@ export default function CustomerTable() {
                                 <td >{customer.lname}</td>
                                 <td >{customer.email}</td>
                                 <td >{customer.address}</td>
-                                <td >{customer.pNo}</td>
+                                <td >{"0"+customer.pNo}</td>
                                 <td >
                                     <Link className="btn ash-btn my-btn"
-                                    to={`/updateCustomer/${customer._id}`} 
+                                    to={`/updateCustomer/${customer.cus_id}`} 
                                     ><svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" className="bi bi-pencil-square" viewBox="0 0 16 16">
                                     <path d="M15.502 1.94a.5.5 0 0 1 0 .706L14.459 3.69l-2-2L13.502.646a.5.5 0 0 1 .707 0l1.293 1.293zm-1.75 2.456l-2-2L4.939 9.21a.5.5 0 0 0-.121.196l-.805 2.414a.25.25 0 0 0 .316.316l2.414-.805a.5.5 0 0 0 .196-.12l6.813-6.814z"/>
                                     <path fillRule="evenodd" d="M1 13.5A1.5 1.5 0 0 0 2.5 15h11a1.5 1.5 0 0 0 1.5-1.5v-6a.5.5 0 0 0-1 0v6a.5.5 0 0 1-.5.5h-11a.5.5 0 0 1-.5-.5v-11a.5.5 0 0 1 .5-.5H9a.5.5 0 0 0 0-1H2.5A1.5 1.5 0 0 0 1 2.5v11z"/>
